Settings: Add support for Lockscreen clock fonts
After r16, a commit to cleanup the KeyguardClockSwtich and unify it's structure was merged.
Also, this added a new resource into the frameworks core, which allows us to overlay the custom clock font.
So, we need to overlays now to customize this. This makes me really happy because it removes the ugly method that we were doing years ago.

[nurkeinneid]:
- adapt to our ThemeUtils
- restore original copyright header - crdroid did nothing here

Change-Id: I70590910144141a534b926f884e9f08e4e3af5bc
Signed-off-by: NurKeinNeid <mralexman3000@gmail.com>
Signed-off-by: aswin7469 <aswinas@pixysos.com>